+++ src/ser.rs
@@ -1,5 +1,5 @@
-use ethers_core::types::Signature;
-use serde::{Deserialize, Serialize};
+use ethers_core::types::{Signature, H160};
+use serde::{Deserialize, Serialize, Serializer};
 
 #[derive(Debug, Clone, Copy, PartialEq)]
 /// Wrapper around a signature that ensures it serializes/deserializes
@@ -32,6 +32,15 @@ impl From<RsvSignature> for Signature {
     }
 }
 
+impl Serialize for RsvSignature {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        serializer.serialize_str(&format!("0x{}", self.0))
+    }
+}
+
 impl<'de> Deserialize<'de> for RsvSignature {
     fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
     where
@@ -44,20 +53,74 @@ impl<'de> Deserialize<'de> for RsvSignature {
     }
 }
 
-impl Serialize for RsvSignature {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+pub(crate) fn serialize_checksum_addr<S>(val: &H160, serializer: S) -> Result<S::Ok, S::Error>
+where
+    S: Serializer,
+{
+    serializer.serialize_str(&ethers_core::utils::to_checksum(val, None))
+}
+
+pub(crate) mod decimal_u64_ser {
+    use ethers_core::types::U64;
+    use serde::{Deserialize, Deserializer, Serializer};
+
+    pub(crate) fn serialize<S>(val: &U64, serializer: S) -> Result<S::Ok, S::Error>
     where
-        S: serde::Serializer,
+        S: Serializer,
     {
-        serializer.serialize_str(&format!("0x{}", self.0))
+        serializer.serialize_str(&val.to_string())
+    }
+
+    pub(crate) fn deserialize<'de, D>(deserializer: D) -> Result<U64, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let s = String::deserialize(deserializer)?;
+        U64::from_dec_str(&s).map_err(serde::de::Error::custom)
+    }
+}
+
+pub(crate) mod json_u256_ser {
+    use ethers_core::types::U256;
+    use serde::{Deserialize, Deserializer, Serialize, Serializer};
+
+    #[derive(Debug, Copy, Clone, Serialize, Deserialize)]
+    struct JsonU256 {
+        hex: U256,
+    }
+
+    pub(crate) fn serialize<S>(val: &U256, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        (JsonU256 { hex: *val }).serialize(serializer)
+    }
+
+    pub(crate) fn deserialize<'de, D>(deserializer: D) -> Result<U256, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        JsonU256::deserialize(deserializer).map(|val| val.hex)
     }
 }
 
 #[cfg(test)]
 mod test {
+    use ethers::prelude::U64;
     use ethers_signers::{LocalWallet, Signer};
 
     use super::*;
+
+    #[test]
+    fn u64_ser() {
+        #[derive(Serialize, Deserialize, Debug)]
+        struct TestU64(#[serde(with = "super::decimal_u64_ser")] U64);
+
+        assert_eq!(
+            "382345198",
+            ethers_core::types::U64::from(382345198).to_string()
+        );
+    }
 
     #[tokio::test]
     async fn sig_serialization() {